Limit stft sizes in cpu version as well

+++ tfr/stft.cpp
@@ -135,7 +135,7 @@ float Fft::
 unsigned Fft::
         next_good_size( unsigned current_valid_samples_per_chunk, float /*sample_rate*/ )
 {
-    size_t maxsize = std::min( (size_t)(64<<20), availableMemoryForSingleAllocation() );
+    size_t maxsize = std::min( (size_t)(64<<20), (size_t)availableMemoryForSingleAllocation() );
     maxsize /= 3*sizeof(ChunkElement);
     return std::min(maxsize, (size_t)Fft::sChunkSizeG(current_valid_samples_per_chunk));
 }
@@ -144,7 +144,7 @@ unsigned Fft::
 unsigned Fft::
         prev_good_size( unsigned current_valid_samples_per_chunk, float /*sample_rate*/ )
 {
-    size_t maxsize = std::min( (size_t)(64<<20), availableMemoryForSingleAllocation() );
+    size_t maxsize = std::min( (size_t)(64<<20), (size_t)availableMemoryForSingleAllocation() );
     maxsize /= 3*sizeof(ChunkElement);
     return std::min(maxsize, (size_t)Fft::lChunkSizeS(current_valid_samples_per_chunk));
 }
@@ -516,7 +516,7 @@ unsigned Stft::
     if (current_valid_samples_per_chunk<_window_size)
         return _window_size;
 
-    size_t maxsize = std::min( (size_t)(64<<20), availableMemoryForSingleAllocation() );
+    size_t maxsize = std::min( (size_t)(64<<20), (size_t)availableMemoryForSingleAllocation() );
     maxsize = std::max((size_t)_window_size, maxsize/(3*sizeof(ChunkElement)));
     return std::min((unsigned)maxsize, spo2g(align_up(current_valid_samples_per_chunk, _window_size)/_window_size)*_window_size);
 }
@@ -528,7 +528,7 @@ unsigned Stft::
     if (current_valid_samples_per_chunk<2*_window_size)
         return _window_size;
 
-    size_t maxsize = std::min( (size_t)(64<<20), availableMemoryForSingleAllocation() );
+    size_t maxsize = std::min( (size_t)(64<<20), (size_t)availableMemoryForSingleAllocation() );
     maxsize = std::max((size_t)_window_size, maxsize/(3*sizeof(ChunkElement)));
     return std::min((unsigned)maxsize, lpo2s(align_up(current_valid_samples_per_chunk, _window_size)/_window_size)*_window_size);
 }
